Ignore specific apps and processes
Added a new rollover effect to each app and process that lets you
toggle whether the icon should animate when it’s using the network.

There are some NASTY hacks in here that need to be cleaned up and
tested/vetted, and it should probably get some usability testing as
well.