Fixing a rendering bug on OS X 10.7
On OS X 10.7 this code would cause menu items to stop rendering once
they were selected for the first time, probably because it actually
used the dummy state we pushed onto the stack. The solution was to make
a copy of the graphics stack before popping it, and push that state
onto the stack when we’re finished custom drawing. It used two more
private and undocumented functions, oh boy!