Update Localizable.strings
I had to replace %@ because it was not showing the 'untranslated' app name, "Loading".

It is better to keep one name for the application in all languages. like all the apps on Mac.

To avoid French nightmare with pluralsssss

I have replaced "loading/loaded" header menu items by Currently and Recently.