Show the checkboxes more often
Only having the checkboxes available when using the mouse and after
holding down the option key made it unintuitive for quite a few people.
Now the checkboxes will be visible when using the keyboard controls and
in the normal menu as well. Hopefully this will be good enough until a
better design comes along!