New Options menu
I had a few suggestions to make it easier to find the website again and
to move the Quit menu to the Options submenu, so this takes care of
both. The About and Send Feedback options were combined into an inline
about menu item, and clicking it takes you to the website.